backend/requirements.txt
# ==============================================================================
# T.A.R.S. Backend Dependencies
# Python 3.11+
# ==============================================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# ASGI Server
uvicorn==0.24.0

# LLM Integration
ollama==0.1.6
langchain==0.1.0
langchain-community==0.0.10

# Vector Database
chromadb==0.4.18
sentence-transformers==2.2.2

# Phase 5: Advanced RAG Dependencies
transformers==4.36.0  # For cross-encoder models
rank-bm25==0.2.2  # BM25 keyword search
langchain-text-splitters==0.0.1  # Semantic text splitting

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# WebSocket Support
websockets==12.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
watchdog==3.0.0
psutil==5.9.6

# Monitoring & Logging
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
pylint==3.0.3
bandit==1.7.5
